% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gafaddin-package.R
\docType{package}
\name{gafaddin}
\alias{gafaddin}
\alias{gafaddin-package}
\title{GitHub Activity Feed RStudio Addin}
\description{
Minimal HTML-view GitHub activity feed viewer.
}
\author{
Bob Rudis (bob@rud.is)
}
